refactored config file reloading
